work around funny variable discrepancy with gmake
In major BSD make(1) implementations (such as the one I'm using with my
FreeBSD box), $> expands to all source files for the target. However, in
GNU make, this variable is $^ instead. I could possibly have just
replaced the $> in my code with $^$> but i think my solution is a little
cleaner and makes it easier to expand on this code in the future. Also
no one is going to know what the hell $^$> means lololol. Maybe
eventually I'll switch to something that uses make suffix rules like
you're supposed to do with makefiles but this seems to work fine for the
way this repo is set up at the moment. I also figure since $^ and $> are
both not portable that there are make implementations that use neither
of those somewhere out there. I understand that in practice people who
don't use GNU make are mostly just restricted to BSD users and idk
people who refuse to install GNU make on their GNU/Linux box for some
reason, but when I code I try to make things as portable as possible
even if that results in slightly goofy scenarios like this. I understand
I could just have made GNU make a requirement for this project or
whatever but I think that misses the point of how I usually think of
writing code like this. Thank  you for coming to my TED talk :P.