how the fuck was that the whole lesson 4
This is probably The End™. I probably won't do lessons 5 or 6 at this
point, I don't really need them anyway. Idk I might reference them if I
need to idk use ttfs in an sdl project later on or smth. Ngl this whole
tutorial was very mid. I can't wait to do an an actual sdl project
instead of slogging through this bs.

Okay to be fair TwinkleBearDev did a very good job at this tutorial and
part of the reason I don't love it is that I'm not really the target
audience for it. I honestly made myself do this so that I could actually
have some practice with SDL again before working on a more serious
project involving it. It was honestly really good for me and I'd say it
did well getting me back into C coding and thinking about documentation.